refactor: 패키지 구조 개선 (Command 응집)
