+++ src/main/java/chess/controller/command/EndCommand.java
@@ -1,21 +1,20 @@
-package chess.controller;
+package chess.controller.command;
 
 import chess.domain.ChessGame;
 import chess.view.OutputView;
 
-public class StartCommand implements Command {
+public class EndCommand implements Command {
     @Override
     public void execute(ChessGame chessGame, OutputView outputView) {
-        outputView.printChessBoardMessage(chessGame.getChessBoard());
     }
 
     @Override
     public boolean isNotEndCommand() {
-        return true;
+        return false;
     }
 
     @Override
     public boolean isNotStartCommand() {
-        return false;
+        return true;
     }
 }
