+++ src/main/java/chess/controller/command/StartCommand.java
@@ -1,20 +1,21 @@
-package chess.controller;
+package chess.controller.command;
 
 import chess.domain.ChessGame;
 import chess.view.OutputView;
 
-public class EndCommand implements Command {
+public class StartCommand implements Command {
     @Override
     public void execute(ChessGame chessGame, OutputView outputView) {
+        outputView.printChessBoardMessage(chessGame.getChessBoard());
     }
 
     @Override
     public boolean isNotEndCommand() {
-        return false;
+        return true;
     }
 
     @Override
     public boolean isNotStartCommand() {
-        return true;
+        return false;
     }
 }
